integrating bulk scan v2
