+++ build.gradle
@@ -84,14 +84,6 @@ sourceSets {
         }
         resources.srcDir file('src/contractTest/resources')
     }
-    smokeTest {
-        java {
-            compileClasspath += main.output
-            runtimeClasspath += main.output
-            srcDir file('src/smokeTest/java')
-        }
-        resources.srcDir file('src/smokeTest/resources')
-    }
 }
 
 dependencies {
