+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/bulkscan/controllers/OcrValidationTest.java
@@ -20,10 +20,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.authorisation.validators.AuthTokenValidator;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
@@ -40,16 +40,16 @@ public class OcrValidationTest  {
     @Autowired
     private MockMvc mvc;
 
-    @MockBean
+    @MockitoBean
     protected AuthTokenValidator authTokenValidator;
 
-    @MockBean
+    @MockitoBean
     protected IdamService idamService;
 
-    @MockBean
+    @MockitoBean
     protected CcdService ccdService;
 
-    @MockBean
+    @MockitoBean
     private RefDataService refDataService;
 
     @Test
