+++ src/functionalTest/java/uk/gov/hmcts/reform/sscs/bulkscan/TransformationFunctionalTest.java
@@ -16,7 +16,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
-import uk.gov.hmcts.reform.sscs.bulkscan.functional.BaseFunctionalTest;
 
 @SpringBootTest
 @TestPropertySource(locations = "classpath:application_e2e.yaml")
