+++ src/functionalTest/resources/smokeRecord.json
