--- src/main/java/uk/gov/hmcts/reform/sscs/bulkscan/bulkscancore/domain/CaseDetails.java
@@ -1,25 +1,0 @@
-package uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.Map;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class CaseDetails {
-
-    @JsonProperty("case_data")
-    private Map<String, Object> caseData;
-
-    @JsonProperty("id")
-    private String caseId;
-
-    private String state;
-}
