+++ src/main/java/uk/gov/hmcts/reform/sscs/bulkscan/controllers/TransformationController.java
@@ -8,21 +8,21 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.reform.sscs.bulkscan.auth.AuthService;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.ExceptionRecord;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.handlers.CcdCallbackHandler;
 import uk.gov.hmcts.reform.sscs.bulkscan.domain.transformation.SuccessfulTransformationResponse;
+import uk.gov.hmcts.reform.sscs.service.AuthorisationService;
 
 @RestController
 public class TransformationController {
 
     private static final Logger LOGGER = getLogger(TransformationController.class);
 
-    private final AuthService authService;
+    private final AuthorisationService authService;
     private final CcdCallbackHandler handler;
 
     public TransformationController(
-        AuthService authService,
+        AuthorisationService authService,
         CcdCallbackHandler handler
     ) {
         this.authService = authService;
