+++ src/main/java/uk/gov/hmcts/reform/sscs/controller/RootController.java
@@ -1,6 +1,9 @@
 package uk.gov.hmcts.reform.sscs.controller;
 
+import static org.springframework.http.ResponseEntity.ok;
+
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -15,6 +18,20 @@ public class RootController {
     @GetMapping(value = "/", consumes = MediaType.ALL_VALUE)
     public void root() {
         // Only used for returning a 200 on /
+    }
+
+    /**
+     * Root GET endpoint.
+     *
+     * <p>Azure application service has a hidden feature of making requests to root endpoint when
+     * "Always On" is turned on.
+     * This is the endpoint to deal with that and therefore silence the unnecessary 404s as a response code.
+     *
+     * @return Welcome message from the service.
+     */
+    @GetMapping
+    public ResponseEntity<String> welcome() {
+        return ok("Welcome to sscs-tribunals-api");
     }
 }
