+++ src/main/java/uk/gov/hmcts/reform/sscs/domain/CaseResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain;
+package uk.gov.hmcts.reform.sscs.domain;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
