--- src/smokeTest/resources/application_smoke.yaml
@@ -1,1 +1,0 @@
-test_url: ${TEST_URL:http://localhost:8090}
