+++ src/test/java/uk/gov/hmcts/reform/sscs/bulkscan/validators/FormTypeValidatorTest.java
@@ -8,17 +8,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.junit.Test;
-import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.CaseResponse;
+import uk.gov.hmcts.reform.sscs.domain.CaseResponse;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.ExceptionRecord;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.OcrDataField;
 import uk.gov.hmcts.reform.sscs.ccd.domain.FormType;
 import uk.gov.hmcts.reform.sscs.bulkscan.json.SscsJsonExtractor;
 
-
-
 public class FormTypeValidatorTest {
-
-
     SscsJsonExtractor sscsJsonExtractor = new SscsJsonExtractor();
     FormTypeValidator validator = new FormTypeValidator(sscsJsonExtractor);
