+++ src/test/java/uk/gov/hmcts/reform/sscs/bulkscan/validators/SscsCaseValidatorTest.java
@@ -56,8 +56,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 import org.springframework.test.util.ReflectionTestUtils;
-import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.CaseDetails;
-import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.CaseResponse;
+import uk.gov.hmcts.reform.sscs.domain.CaseDetails;
+import uk.gov.hmcts.reform.sscs.domain.CaseResponse;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.ExceptionRecord;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.OcrDataField;
 import uk.gov.hmcts.reform.sscs.bulkscan.bulkscancore.domain.ScannedData;
